fix commands format
