+++ README.md
@@ -18,7 +18,7 @@ go get github.com/jmcvetta/napping
 ```
 
 ### 環境設定
-LBのエンドポイントを追加
+EC2のプライベートDNSを追加
 
 ```
 vi ~/.bashrc
@@ -28,6 +28,11 @@ export WCAFE_VM_PRIVATE_DNS=private_dns
 source ~/.bashrc
 ```
 
+### 動作確認
+```
+go run main.go stores list
+```
+
 ### コマンドのインストール
 ```
 go build
@@ -35,15 +40,15 @@ go install
 ```
 
 ### 使い方
-#### getの場合
+#### 一覧データの取得の場合
 ```
-wcafe-cli get stores
-wcafe-cli get pets
-wcafe-cli get users
+wcafe-cli stores list
+wcafe-cli pets　list
+wcafe-cli users　list
 ```
-#### postの場合
+#### 新規作成の場合
 ```
-wcafe-cli post stores
-wcafe-cli post pets <store_id>
-wcafe-cli post users
+wcafe-cli stores　create
+wcafe-cli pets  create <store_id>
+wcafe-cli users create
 ```
